Bug 721888 - Fix compiler warnings. r=blassey
UltraBlame original commit: 2571f70b20c5c60b54c2452e41b7a172280778a4

+++ mobile/android/base/GeckoEvent.java
@@ -193,7 +193,7 @@ public class GeckoEvent {
         PointF geckoPoint = new PointF(event.getX(eventIndex), event.getY(eventIndex));
         geckoPoint = GeckoApp.mAppContext.getLayerController().convertViewPointToLayerPoint(geckoPoint);
 
-        mPoints[index] = new Point((int)Math.round(geckoPoint.x), (int)Math.round(geckoPoint.y));
+        mPoints[index] = new Point(Math.round(geckoPoint.x), Math.round(geckoPoint.y));
         mPointIndicies[index] = event.getPointerId(eventIndex);
         
         if (Build.VERSION.SDK_INT >= 9) {
